css and new func
created new func (completion time, task priority, dummy task for add new todo), styled css, tested all the things, bugfixed. finishing touches will come, but not necessary

+++ dp-app-fe/src/TodoItem/TodoItem.js
@@ -6,28 +6,36 @@ class TodoItem extends React.Component{
     render(){
 
         const todo = this.props.todo;
+        const items = ["left", "midleft", "middle", "midright", "right"];
         return(
-            <div className = {'todoContainer'}>
-                <form onSubmit={(e) => this.deleteTodo(e)}>
-
-                    <div className={'dueDate'}>Finish till: {todo.dueDate}</div>
+            <div className = {'speech-bubble ' + items[Math.floor(Math.random()*items.length)]}>
+                <br />
+                <form>
                     <div className={'title' + (todo.finished ? ' finished' : '')} onClick={this.markFinished}>{todo.title}</div>
                     <div className={'task' + (todo.finished ? ' finished' : '')} onClick={this.markFinished}>{todo.text}</div>
+                    <button className='deleteButton' id={'finish'} onClick={(e) => this.deleteDisableTodo(e)}>Finish</button>
+                    <button className='deleteButton' id={'delete'} onClick={(e) => this.deleteDisableTodo(e)}>Delete</button>
+                    <div className={'dueDate'}>Finish till: {todo.dueDate}</div>
+                    {todo.priority!=='' ? (<div className ={'priority'}>Priority: {todo.priority}</div>) : (<p></p>)}
                     <div className ={'dateOfCreation'}>Created: {todo.createdAt}</div>
-
-                    <button className='deleteButton' type={'submit'}>Delete</button>
+                    {todo.finished===true ? (<div className ={'dateOfCompletion'}>Completed: {todo.completedAt}</div>) : (<p></p>)}
                 </form>
+                <br />
             </div>
         );
     }
 
     markFinished = () =>{
         this.props.updateTodoFn(this.props.todo);
+
     };
 
-    deleteTodo = (e) => {
-        console.log('delete');
-        this.props.deleteTaskFn(this.props.todo);
+    deleteDisableTodo = (e) => {
+        if (e.target.id === 'delete') {
+            this.props.deleteTaskFn(this.props.todo);
+        }else{
+            this.markFinished();
+        }
     };
 
 }
